pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "surgical-segmentation"
version = "0.1.0"
description = "Deep learning pipeline for surgical instrument segmentation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Maximilian Herbert Dressler", email = "maximilian.dressler@stud.uni-heidelberg.de"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
keywords = ["surgical", "segmentation", "deep learning", "medical imaging", "laparoscopy"]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
jupyter = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipykernel>=6.25.0",
]

[project.urls]
Homepage = "https://github.com/Herbert-Research/surgical-instrument-segmentation"
Documentation = "https://github.com/Herbert-Research/surgical-instrument-segmentation#readme"
Repository = "https://github.com/Herbert-Research/surgical-instrument-segmentation.git"
Issues = "https://github.com/Herbert-Research/surgical-instrument-segmentation/issues"

[project.scripts]
train-segmentation = "surgical_segmentation.training.trainer:main"
evaluate-model = "surgical_segmentation.evaluation.analyzer:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=surgical_segmentation --cov-report=term-missing"
